refactor: split mcp.ts into focused modules
Split the 1,317-line mcp.ts into 5 focused modules:

- mcp-types.ts          — shared types, protocol constants, isRecord helper
- mcp-tool-naming.ts    — collision-safe tool name allocation utilities
- mcp-transport-stdio.ts — STDIO transport (child-process MCP servers)
- mcp-transport-http.ts  — HTTP + Streamable HTTP transports (incl. OAuth)
- mcp-registry.ts        — client factory, OAuth lifecycle, tool registration

mcp.ts becomes a barrel re-export so all existing callers (mcp-extension.ts,
mcp.test.ts) continue to import from ./mcp.js with no changes needed.

No logic changes — pure reorganization only. All 23 tests pass.
